Finish styling. Add mousedown operation.
Styling finished and elements re-jigged, mainly to keep the grid and header
central.
Added extra onMouseDown and onMouseUp features to the gridContainer so that
the color only changes with mouseover and mousedown together (click and drag
to draw). Default behavior also disabled on mousedown within the comtainer,
otherwise it was possible to start dragging and mess the thing up.